Refactor: main 폴더 생성, 파일 분리하기
