+++ client/front/lib/errands/fixerrand/fixerrandwidget/fixminimap.dart
