+++ client/front/lib/main/writeerrand/errand_check.dart
@@ -6,11 +6,11 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:front/errands/fixerrand/fixerrandwidget/fixerrand.dart';
 import 'package:intl/intl.dart';
-import 'fixerrand/fixerrandwidget/fixerrand.dart';
-import '../mainshowerrand/mainshowerrand.dart';
+import '../../mainshowerrand/mainshowerrand.dart';
 import 'request.dart';
-import 'main_post_page.dart';
+import '../home/main_post_page.dart';
 import 'package:http/http.dart' as http;
 
 final storage = FlutterSecureStorage(); // 토큰 받기
