+++ client/front/lib/mainshowerrand/mainshowerrand.dart
@@ -1,15 +1,15 @@
 import 'dart:convert';
 import 'dart:developer';
-import 'package:front/errand/errand_check.dart';
+import 'package:front/main/writeerrand/errand_check.dart';
 import 'package:http/http.dart' as http;
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:front/errand/main_post_page.dart';
+import 'package:front/main/home/main_post_page.dart';
 import 'package:front/mainshowerrand/showerrandwidget/showerrandwidget.dart';
-import 'package:front/errand/request.dart';
+import 'package:front/main/writeerrand/request.dart';
 
 final storage = FlutterSecureStorage(); // 토큰 받기
