+++ client/front/lib/mainshowerrand/showerrandwidget/showerrandwidget.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:front/errand/errand_check.dart';
+import 'package:front/main/writeerrand/errand_check.dart';
 import 'package:front/mainshowerrand/showerrandwidget/stamp.dart';
 import 'package:front/mainshowerrand/showerrandwidget/tablescreen/tablescreen1.dart';
 import 'package:front/mainshowerrand/showerrandwidget/tablescreen/tablescreen2.dart';
